feat: 엔티티와 순수 함수 분리를 기준으로 새로운 폴더 계층 구조 변경 1. compmonents 도메인 별 분리
