--- src/basic/components/features/index.ts
@@ -1,6 +1,0 @@
-// Feature Components (기능별 중간 레벨 컴포넌트들)
-
-export { Header } from "./Header";
-export { ProductList } from "./ProductList";
-export { CartSidebar } from "./CartSidebar";
-export { AdminPanel } from "./AdminPanel";
