+++ src/basic/components/index.ts
@@ -1,5 +1,18 @@
-// UI Components (기본 재사용 단위)
-export * from "./ui";
+/**
+ * Components - 새로운 도메인 구조로 re-export
+ */
 
-// Feature Components (기능별 중간 레벨)
-export * from "./features";
+// Shared components (migrated to shared/components)
+export * from "../shared/components";
+
+// Product domain components (migrated to domain/product/components)
+export * from "../domain/product/components";
+
+// Cart domain components (migrated to domain/cart/components)
+export * from "../domain/cart/components";
+
+// Notification domain components (migrated to domain/notification/components)
+export * from "../domain/notification/components";
+
+// Feature components (migrated to features/)
+export * from "../features";
