--- src/basic/components/ui/index.ts
@@ -1,4 +1,0 @@
-// UI Components (재사용 가능한 기본 단위)
-export { NotificationToast } from "./NotificationToast";
-export { SearchInput } from "./SearchInput";
-export { ProductCard } from "./ProductCard";
