+++ src/basic/domain/cart/components/index.ts
@@ -1,0 +1,2 @@
+// Cart domain components
+export { CartSidebar } from "./CartSidebar";
