+++ src/basic/domain/notification/components/NotificationToast.tsx
@@ -1,0 +1,56 @@
+import React from "react";
+import { Notification } from "../models";
+
+interface NotificationToastProps {
+  notifications: Notification[];
+  onRemove: (id: string) => void;
+}
+
+export const NotificationToast: React.FC<NotificationToastProps> = ({
+  notifications,
+  onRemove,
+}) => {
+  if (notifications.length === 0) return null;
+
+  return (
+    <div className="fixed top-4 right-4 z-50 space-y-2">
+      {notifications.map((notification) => (
+        <div
+          key={notification.id}
+          className={`
+            max-w-sm rounded-lg shadow-lg p-4 transition-all duration-300 transform
+            ${
+              notification.type === "error"
+                ? "bg-red-500 text-white"
+                : notification.type === "warning"
+                ? "bg-yellow-500 text-white"
+                : "bg-green-500 text-white"
+            }
+          `}
+        >
+          <div className="flex justify-between items-start">
+            <p className="text-sm font-medium">{notification.message}</p>
+            <button
+              onClick={() => onRemove(notification.id)}
+              className="ml-4 text-white hover:text-gray-200 transition-colors"
+            >
+              <svg
+                className="w-4 h-4"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M6 18L18 6M6 6l12 12"
+                />
+              </svg>
+            </button>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+};
