+++ src/basic/domain/notification/components/index.ts
@@ -1,0 +1,2 @@
+// Notification domain components
+export { NotificationToast } from "./NotificationToast";
