+++ src/basic/domain/product/components/ProductList/index.ts
