+++ src/basic/domain/product/components/index.ts
@@ -1,0 +1,3 @@
+// Product domain components
+export { ProductCard } from "./ProductCard";
+export { ProductList } from "./ProductList";
