+++ src/basic/features/admin/index.ts
@@ -1,0 +1,2 @@
+// Admin feature components - 여러 도메인을 조합한 관리자 기능
+export { AdminPanel } from "./AdminPanel";
