+++ src/basic/features/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Features - 여러 도메인을 조합한 복합 기능 컴포넌트들
+ */
+
+// Admin features
+export * from "./admin";
