+++ src/basic/shared/components/form/index.ts
@@ -1,0 +1,2 @@
+// Form components - 범용 폼 관련 컴포넌트들
+export { SearchInput } from "./SearchInput";
