+++ src/basic/shared/components/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared components - 엔티티와 무관한 범용 컴포넌트들
+ */
+
+// Layout components
+export * from "./layout";
+
+// Form components
+export * from "./form";
