+++ src/basic/shared/components/layout/index.ts
@@ -1,0 +1,2 @@
+// Layout components - 범용 레이아웃 컴포넌트들
+export { Header } from "./Header";
